requirements.txt
langchain
beautifulsoup4
tiktoken
openai
pinecone-client
unstructured
nltk
python-magic-bin
fastapi
jinja2
uvicorn
streamlit
streamlit-chat
tqdm
python-dotenv
google-search-results
faiss-cpu
lxml
sentence-transformers